Move set avatar speed to the animation reactor
