+++ packages/engine/src/avatar/functions/avatarFunctions.ts
@@ -24,22 +24,19 @@ Infinite Reality Engine. All Rights Reserved.
 */
 
 import { VRM, VRMHumanBone, VRMHumanBoneList } from '@pixiv/three-vrm'
-import { AnimationClip, AnimationMixer, Matrix4, Vector3 } from 'three'
+import { AnimationMixer, Matrix4, Vector3 } from 'three'
 
 import { getComponent, getOptionalComponent, hasComponent, setComponent } from '@ir-engine/ecs/src/ComponentFunctions'
 import { Entity } from '@ir-engine/ecs/src/Entity'
-import { getMutableState, getState } from '@ir-engine/hyperflux'
+import { getState } from '@ir-engine/hyperflux'
 import { iterateEntityNode } from '@ir-engine/spatial/src/transform/components/EntityTree'
 import { TransformComponent } from '@ir-engine/spatial/src/transform/components/TransformComponent'
 import { computeTransformMatrix } from '@ir-engine/spatial/src/transform/systems/TransformSystem'
 
 import { AnimationState } from '../AnimationManager'
-import { getRootSpeed } from '../animation/AvatarAnimationGraph'
-import { preloadedAnimations } from '../animation/Util'
 import { AnimationComponent } from '../components/AnimationComponent'
 import { AvatarRigComponent } from '../components/AvatarAnimationComponent'
 import { AvatarComponent } from '../components/AvatarComponent'
-import { AvatarMovementSettingsState } from '../state/AvatarMovementSettingsState'
 
 declare module '@pixiv/three-vrm/types/VRM' {
   export interface VRM {
@@ -116,28 +113,6 @@ export const setAvatarAnimations = (entity: Entity) => {
       .flat(),
     mixer: new AnimationMixer(vrm.humanoid.normalizedHumanBonesRoot)
   })
-}
-
-const runClipName = 'Run_RootMotion',
-  walkClipName = 'Walk_RootMotion'
-
-/**
- * @todo: stop using global state for avatar speed
- * in future this will be derrived from the actual root motion of a
- * given avatar's locomotion animations
- */
-export const setAvatarSpeedFromRootMotion = () => {
-  const manager = getState(AnimationState)
-  const animations = getComponent(
-    manager.loadedAnimations[preloadedAnimations.locomotion],
-    AnimationComponent
-  ).animations
-  /**@todo handle avatar animation clips generically */
-  const run = AnimationClip.findByName(animations, runClipName)
-  const walk = AnimationClip.findByName(animations, walkClipName)
-  const movement = getMutableState(AvatarMovementSettingsState)
-  if (run) movement.runSpeed.set(getRootSpeed(run))
-  if (walk) movement.walkSpeed.set(getRootSpeed(walk))
 }
 
 export const getAvatarBoneWorldPosition = (entity: Entity, boneName: string, position: Vector3): boolean => {
